Add files via upload
get_initial_tree method is now working! It returns an array of the tasks that should be assigned to the machines (arr[k] = machine (k+1)'s task where A=0, B=1, ... , H=7)
get_cost method is now working! (returns an integer value for the cost with -1 => machine task pairing is not valid)